librad: reorganise integration tests
The slightly off module structure is due to cargo compling all top-level
modules into executables.

There are certainly more opportunities to extract common subroutines,
but this is left to a future brooming pass.
