+++ ci/run
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 set -eou pipefail
 
-./ci/build-test
-./ci/advisory
 ./ci/clippy
+./ci/build-test
 ./ci/docs
+./ci/advisory
